backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud Services
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.36.4
google-cloud-translate==3.12.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
python-magic-bin==0.4.14  # For Windows

# Vector & AI
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Environment
python-dotenv==1.0.0